Split into internal and external JsonFeed data structures
The internal one knows all fields across spec versions and the one
exported as part of the public API always points to the latest spec
revision. Everything will be converted to the new one.

For example, if you parse a JsonFeed 1 feed with this and then export
it, it will then be in valid JsonFeed 1.1 format.